* Fixes * creature speak type changable - doCreatureSetSpeakType(cid, speakType) * Saving player namelocks in database * Replaced doCreatureSetNameDescription with doCreatureSetSpecialDescription (It couldn't be usable) * New functions: setCreatureMaxCap, getTalkactionsList, getExperienceStagesList * Invisible GM won't show effect while casting spells (configurable) * Items limit per tile (House & protection tile) * Sad way for yalahar stairs (But it works same as in global tibia) * Configurable disabled mailbox ids (For example, you can block rook) * More.. [?]
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@2173 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ vocation.cpp
@@ -347,7 +347,7 @@ uint64_t Vocation::getReqMana(uint32_t magLevel)
 	if(it != cacheMana.end())
 		return it->second;
 
-	uint64_t reqMana = (uint64_t)(400 * std::pow(formulaMultipliers[MULTIPLIER_MANA], (magLevel - 1)));
+	uint64_t reqMana = (uint64_t)(400 * std::pow(formulaMultipliers[MULTIPLIER_MANA], (float)(magLevel - 1)));
 	if(reqMana % 20 < 10)
 		reqMana -= reqMana % 20;
 	else
